+++ api-site/modules/library/domain/map_asset_aggregate.py
@@ -34,7 +34,7 @@ class MapAsset(MediaAssetAggregate):
     grid_offset_x: Optional[int] = None
     grid_offset_y: Optional[int] = None
     grid_line_color: Optional[str] = None
-    grid_cell_size: Optional[int] = None
+    grid_cell_size: Optional[float] = None
 
     @classmethod
     def create(
@@ -87,7 +87,7 @@ class MapAsset(MediaAssetAggregate):
         grid_offset_x: Optional[int] = None,
         grid_offset_y: Optional[int] = None,
         grid_line_color: Optional[str] = None,
-        grid_cell_size: Optional[int] = None
+        grid_cell_size: Optional[float] = None
     ) -> "MapAsset":
         """
         Promote a base MediaAssetAggregate to MapAsset.
@@ -122,7 +122,7 @@ class MapAsset(MediaAssetAggregate):
         grid_offset_x: Optional[int] = None,
         grid_offset_y: Optional[int] = None,
         grid_line_color: Optional[str] = None,
-        grid_cell_size: Optional[int] = None
+        grid_cell_size: Optional[float] = None
     ) -> None:
         """
         Update grid configuration.
